% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-package.R
\docType{package}
\name{cellTypeCompositions-package}
\alias{cellTypeCompositions}
\alias{cellTypeCompositions-package}
\title{cellTypeCompositions: Cell Type Compositions Gibbs Sampler}
\description{
Estimate cell type compositions from a table of clone
      by cell type counts. Functions are provided for exploring the cell
      type compositions that are displayed by cell lineages.  The
      workhorse is a function that performs blocked Gibbs sampling on the
      cell type frequencies in a parent population under a model that
      assumes a vague prior for the total number of cells and a uniform
      Dirichlet prior for the proportions fo different cell types.
}
\details{
The only functions you're likely to need from cellTypeCompositions is 
\code{gibbsScan} and its \code{update} method
}
\keyword{package}
