% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbsDPP.R
\name{gibbsDPP}
\alias{gibbsDPP}
\alias{gibbsScan}
\alias{update.ctScan}
\alias{keep.default}
\title{Dirichlet Process Prior Gibbs Sampler}
\usage{
gibbsDPP(
  wtab,
  om,
  eta = NULL,
  etaN = NULL,
  dataToEta = NULL,
  lambda = NULL,
  lambdaN = NULL,
  dataToLambda = NULL,
  etaM = 0L,
  auxEtaM = 5L,
  alphaEta = 100,
  lambdaM = 0L,
  auxLambdaM = 5L,
  alphaLambda = 1,
  verbose = 0L,
  etaCols = 500L,
  lambdaSize = 10L,
  dprior = 1,
  lambdaShape = 1,
  lambdaRate = 0.01,
  ijvals = 0L,
  ...
)

gibbsScan(
  wtab,
  om,
  eta = NULL,
  etaN = NULL,
  dataToEta = NULL,
  etaM = 0L,
  alphaEta = 1,
  auxEtaM = 10L,
  lambda = NULL,
  lambdaN = NULL,
  dataToLambda = NULL,
  lambdaM = 0L,
  alphaLambda = 1,
  auxLambdaM = 10L,
  nkeep = 1L,
  nthin = 1L,
  nburn = 0L,
  etaCols = 50L,
  lambdaSize = 20L,
  ijvals = 0L,
  abEta = c(1e-04, 1e-04),
  abLambda = c(0.01, 0.01),
  verbose = FALSE,
  dprior = 1,
  lambdaShape = 1,
  lambdaRate = 0.01,
  keep = TRUE,
  niter.tune = 10L,
  ...
)

\method{update}{ctScan}(object, elt = length(object), ...)

keep.default(pass, log.posterior, i, nkeep)
}
\arguments{
\item{wtab}{the result of \code{\link{uniTab}}}

\item{om}{the object returned by \code{\link{omega}(paramList)}}

\item{eta}{(optional) initialization and workspace for the sampled
parameters}

\item{etaN}{(optional) initialization and workspace for the counts
of sampled parameter vectors}

\item{dataToEta}{index from the original data to the value of eta
sampled}

\item{lambda}{(optional) initialization and workspace for the
sampled abundance parameters}

\item{lambdaN}{(optional) initialization and workspace for the
counts of sampled parameters}

\item{dataToLambda}{index from the original data to the value of
lambda sampled}

\item{etaM}{number of samples provided in the intialization}

\item{auxEtaM}{how many auxiliary paramters to draw}

\item{alphaEta}{how much weight to place on new samples}

\item{lambdaM}{how many auxiliary paramters to draw}

\item{auxLambdaM}{how much space to allot for new samples. Only
one new sample will be drawn if greater than zero}

\item{alphaLambda}{how much weight to place on new samples}

\item{verbose}{whether to report intermediate actions}

\item{etaCols}{how large to make the workspace (too small a value
will result in termination with an error)}

\item{lambdaSize}{dprior prior value for draws of eta}

\item{dprior}{symmetric Dirichlet prior value}

\item{lambdaShape}{base gamma distribution prior shape}

\item{lambdaRate}{base distribution prior rate}

\item{ijvals}{\code{0L} by default, but for (not yet implemented)
splitting and merging a value of 2 is needed. Users should not
usually change the default.}

\item{...}{parameters to revise using \code{update}}

\item{nkeep}{the maximum number of iterations}

\item{nthin}{how many iterations per saved value}

\item{nburn}{discard this many iterations before saving any values}

\item{abEta}{if \code{NULL} treat \code{alphaEta} as a fixed
value. Otherwise the first two elements are the shape and
rate parameters of Gamma prior for \code{alphaEta}.}

\item{abLambda}{akin to \code{abEta}}

\item{keep}{either \code{TRUE} to apply \code{keep.default} to the
result in each iteration or a function called with
\code{pass}, the intermediate results, and the other arguments
of \code{keep.default}}

\item{niter.tune}{how many cycles of tuning for each gibbs scan}

\item{object}{a \dQuote{ctScan} object}

\item{elt}{integer to select a sample to use as the starting point}

\item{pass}{a list of parameter values following an iteration, viz.
the values in elements of the default return value of \code{gibbsScan}.}

\item{log.posterior}{the loglikelihood evaluated at the current
values of \code{pass}}

\item{i}{the iteration number}
}
\value{
list which by default has elements \code{eta}, \code{etaN}, \code{etaM},
    \code{lambda}, \code{lambdaN}, \code{lambdaM}, and
    \code{dataToEta} and \code{dataToLambda} which are updates to
    the correspondingly named inputs for \code{gibbsDPP} and a
    list of lists of such elements for \code{gibbsScan}. The
    element \code{logLik} is the loglikelihood for \code{lambda}
    and \code{eta} conditioning on the values of all other
    parameters.
}
\description{
Auxiliary Gibbs Sampler for Dirichlet Process Prior
}
\details{
The auxiliary Gibbs sampler for a Dirichlet Process prior is
implemented. The usual call is to \code{gibbsScan}, but
\code{update} can be used to continue a chain (with no arguments)
possibly changing the parameters.

\code{gibbsDPP} only does an initial Gibbs scan with no following
tuning of parameters or subsequent iteration.  It may be helpful
for low-level access to the scanning routine, but rarely is used.

\code{gibbsScan} is the usual entry point.  It conducts the Gibbs
scan for \code{eta} and \code{lambda} and subseqeunt tuning alone the
lines of Algorithm 8 in Neal, 2000, but the scan for \code{lambda}
uses the integrated posterior from the base distribution as in
Algorithm 1 of Neal, 2000 to decide whether to add a new value.

The method for \code{update} allows iterations under
\code{gibbsScan(..., keep = TRUE)} to resume. If another value is
used for \code{keep} some care is needed to mimic the return value
as described next.

\code{keep.default} is used internally under the default
(\code{gibbsScan(..., keep = TRUE)}) to return values from each
iteration. Users can provide their own function to alter the
collection of values from each iterate. The function should have
the same number of arguments as \code{keep.default} and may pass
those arguments to \code{keep.default} when \code{i == nkeep} to
allow \code{update} to resume iteration.
}
\examples{
tab <- diag(10, nrow = 3)[ rep(1:3,2), ] +
  rbinom(18,2,0.5)
uop <- prop.table(diag(3)+0.05,1)
gibbsScan(uniTab(tab), uop)
}
\references{
Radford M. Neal (2000) \emph{Markov Chain Sampling
    Methods for Dirichlet Process Mixture Models}, Journal of
    Computational and Graphical Statistics, 9:2, 249-265, DOI:
    10.1080/10618600.2000.10474879
}
\author{
Charles Berry
}
