% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R
\name{omega}
\alias{omega}
\title{Expected Observation Counts}
\usage{
omega(pl)
}
\arguments{
\item{pl}{list with elements \code{upsilon}, \code{xi}, and \code{psi}}
}
\value{
matrix of multipliers
}
\description{
Convert Expected Sample Counts to Expected Observation Counts
}
\author{
Charles Berry
}
