% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtab.R
\name{rtab}
\alias{rtab}
\title{rtab}
\usage{
rtab(scan, om, elt = length(scan), tol = 0.001, impute.unseen = TRUE)
}
\arguments{
\item{scan}{the object produced by \code{\link{gibbsScan}} or a
list whose last element has components \dQuote{dataToEta},
\dQuote{dataToLambda}. \dQuote{eta}, and \dQuote{lambda} akin
to those of a \code{ctScan} object.}

\item{om}{a filtration matrix such as made by \code{\link{omega}}}

\item{elt}{which element of the scan to use in sampling.}

\item{tol}{lower limit on probability of observing a clone}

\item{impute.unseen}{if \code{TRUE} unseen clones will be
sampled. This might be \code{FALSE} if \code{scan} represents
a population that is to be sampled.}
}
\value{
table of counts
}
\description{
Sample Cell Type Table
}
\details{
Given the \code{ctScan} fit of a dataset, create a sample table
from the posterior.
}
\examples{
ascan <- list()
ascan$dataToEta <- rep(1:7,each=3)
ascan$dataToLambda <- rep(1:3, 7)
ascan$lambda <- c( 1.0, 5.0, 20.0)
ascan$eta <- cbind( diag(3) * 0.7 + 0.1, 1/3, -0.3 * diag(3)+0.4)
ascan <- list( ascan )
om <-  diag(9:7)/10 + 0.025
atab <- rtab(ascan, om)
}
\author{
Charles Berry
}
