% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniTab.R
\name{uniTab}
\alias{uniTab}
\title{uniTab Combine Duplicate Rows}
\usage{
uniTab(tab, omitNullRows = TRUE)
}
\arguments{
\item{tab}{matrix or table of counts.  Rows correspond to lineages
and columns to cell types.}

\item{omitNullRows}{logical, if \code{TRUE} omit rows with only
zero cells}
}
\value{
\code{list} with elements \code{tab} - the rows of
    \code{unique(tab)},
    \code{n} - a vector of replicate row
    counts,
    \code{data.index} - a mapping of the rows to the
    \code{tab} argument to those of \code{uniTab(tab)[["tab"]]},
    and
    \code{omitted} - \code{NULL} if no rows were omitted, or
    an object such that \code{inverse.rle(result$omitted)} is a
    logical vector indicating which table rows were all zeros.
}
\description{
Combine duplicate rows of a table of counts
}
\details{
When a table of counts has many duplicate rows, computations are
sometimes faster if duplicate rows are combined and a vector of
case weights is created.
}
\author{
Charles Berry
}
